Fix timing issue on ec2 instance creation
Noticed this issue around aws-sdk(-v1) 1.66.0 so it may be a
regression or intentional tightening of expectations, or it may be a
timing issue associated with network latency.

Can reliable reproduce a failure to create an instance, erroring on a
subsequent call to Instance#add_tag because the instance does not
exist (yet).

Avoid this by waiting immediately for the instance to have
status != pending.