Avoid setting Env['GEM_HOME'] with jruby 9.x in installed profile
This was never a good substitute for the jruby.gem.home system
property. It was originally set on 9.x to avoid a constant warning when
the property was set. Instead by default, just take the default
in-dist path of jruby 9.x and system gems will need to be
re-installed with a jruby upgrade.

In addition, a setter for JRubyVM#jruby_gem_home_prop is added to
allow reverting to the old behavior (which will result in the
warnings).