Fix logic bug in Debian#dist_update
Using context state in dist_update no longer worked as expected since
we changed to remote, conditional install
(dist_if_not_installed?). The update state would be set even if the
remote condition was false and there update wasn't really run.

The fix is to change to using a remote lock file where `apt-get
update` is only attempted when (1) dist_if_not_installed? is true and
packages are really going to install and (2) it wasn't already run
within the last hour, based on the lock file which is touched on
success.